+++ R/seriation-helpers.R
@@ -23,6 +23,9 @@ mat_seriate <- function(mat,
 # get list including seriation object, order from that object, and any tree
 rowSeriationInfo <- function(mat, method, dist) {
 
+  # don't try and sort if there's only one row
+  if (nrow(mat) == 1) method <- "Identity"
+
   # get lists of possible methods
   matrixMethods <- seriation::list_seriation_methods(kind = "matrix")
   distanceMethods <- seriation::list_seriation_methods(kind = "dist")